# src/docs/swagger.yaml
basePath: /
definitions:
  errors.ErrorDetails:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        description: ESTE NO IRIA MAS
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  errors.ErrorDetailsWithParams:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
      validation_errors:
        items:
          $ref: '#/definitions/errors.ErrorParam'
        type: array
    type: object
  errors.ErrorParam:
    properties:
      field:
        type: string
      reason:
        type: string
    type: object
...
  models.UserAdditionalRequest:
    properties:
      about_me:
        type: string
      age:
        type: integer
      gender:
        type: string
      location:
        type: string
    type: object
  models.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.UserResponse:
    properties:
      about_me:
        type: string
      age:
        type: integer
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_author:
        type: boolean
      last_name:
        type: string
      location:
        type: string
      username:
        type: string
    type: object
  models.UserStageRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      is_author:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  models.UserStageResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id_register:
        type: string
      is_author:
        type: boolean
      last_name:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a  server for Swagger with Gin.
  title: BetterReads API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      description: Get user by id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get user by id
      tags:
      - users
...
  /users/{id}/picture:
    get:
      description: Get user picture
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get user picture
      tags:
      - users
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Log in a user and return a JWT
      parameters:
      - description: User login request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Log in a user
      tags:
      - users
  /users/picture:
    post:
      consumes:
      - multipart/form-data
      description: Post a picture
      parameters:
      - description: User picture
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Post a picture
      tags:
      - users
  /users/register/{id}/additional-info:
    post:
      consumes:
      - application/json
      description: Register second step
      parameters:
      - description: User register id
        in: path
        name: id
        required: true
        type: string
      - description: User additional request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserAdditionalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Register second step
      tags:
      - users
  /users/register/basic:
    post:
      consumes:
      - application/json
      description: Register first step
      parameters:
      - description: User stage request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserStageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserStageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Register first step
      tags:
      - users
  /users/search:
    get:
      consumes:
      - application/json
      description: Search users
      parameters:
      - description: User name
        in: query
        name: name
        type: string
      - description: Is author
        in: query
        name: author
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Search users
      tags:
      - users
...
swagger: "2.0"